pyproject.toml
[tool.poetry]
name = "oeliks-bot"
version = "0.1.0"
description = ""
authors = ["Nick <nickphilomath@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
telethon = "^1.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
